+++ HISTORY.md
@@ -1,8 +1,9 @@
 ## Changelog ##
 
-### 2014-09-11 - v5.4 - (Dev time: 15 hours) ###
+### 2014-09-11 - v5.4 - (Dev time: 17 hours) ###
 - New: Added the ability to delete post meta fields (Issue #43)
 - New: Added the ability to delete comment meta fields (Issue #70)
+- New: Added the ability to delete user meta fields (Issue #87)
 - New: Added the ability to hook into JavaScript message, date picker and validation (Issue #82, #83, #84)
 - New: Prevent PHP from timing out while performing bulk operations(Issue #81)